+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Features
 
 * Add converter libraries to default libraries
+* Add techdivision/import#162
 * Add techdivision/import-cli-simple#216
 * Add techdivision/import-configuration-jms#25
 * Update to versions of techdivision/import 15.*, techdivision/import-configuration-jms 15.* and techdivision/import-app-simple 16.*